requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2020.12.5
cryptography==39.0.2
dnspython==1.16.0
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
uvicorn==0.20.0
passlib==1.7.4
pyasn1==0.4.8
pymongo==4.6.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.7.2
sniffio==1.3.0